pyproject.toml
[project]
name = "selenium-python-home-task"
description = "Selenium Python example project with pytest and Allure report"
version = "0.1.0"
dependencies = [
  "allure-pytest==2.15.0",
  "assertpy==1.1",
  "pytest==8.4.2",
  "pytest-base-url==2.1.0",
  "pytest-check==2.6.0",
  "pytest-dependency==0.6.0",
  "pytest-ordering==0.6",
  "pytest-rerunfailures==16.1",
  "pytest-split==0.10.0",
  "python-dotenv==1.2.1",
  "requests==2.32.5",
  "selenium==4.38.0",
  "tenacity==9.1.2"
]
requires-python = "~=3.11"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = [
  "--clean-alluredir",
  "--alluredir=allure-results",
  "--allure-link-pattern",
  "issue:https://{}",
  "--allure-link-pattern",
  "link:https://{}",
  "--allure-link-pattern",
  "tms:https://{}",
  "--color=yes",
  "-s",
  "-p no:logging"
]
base_url = 'https://www.twitch.tv/'

testpaths = [
  "tests"
]

[tool.ruff]
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
